openhr20-daemon/requirements.txt
bottle==0.12.23
future==0.18.3
iso8601==1.1.0
paho-mqtt==1.6.1
pyserial==3.5
PyYAML==6.0
websockets==10.4
websocket-client==1.4.2
python-dotenv==0.21.0
pytz==2022.7.1
